+++ ha-integration/custom_components/ha_tv_pip/const.py
@@ -2,6 +2,7 @@
 
 DOMAIN = "ha_tv_pip"
 
+BUTTON_DOMAIN = "button"
 CONF_API_VERSION = "api_version"
 CONF_DEVICE_ID = "device_id"
 CONF_HOST = "host"
@@ -10,8 +11,11 @@ CONF_PAIRING = "pairing"
 CONF_PORT = "port"
 CONF_TOKEN = "token"
 CONF_VERSION = "version"
+SENSOR_DOMAIN = "sensor"
+BINARY_SENSOR_DOMAIN = "binary_sensor"
 
 DEFAULT_PORT = 8765
 DISCOVERY_SERVICE_TYPE = "_ha-tv-pip._tcp.local."
 SERVICE_SHOW_CAMERA = "show_camera"
 SERVICE_SHOW_SNAPSHOT = "show_snapshot"
+PLATFORMS = [SENSOR_DOMAIN, BINARY_SENSOR_DOMAIN, BUTTON_DOMAIN]